loaders/python/requirements.txt
skill-universal>=1.0.0
anthropic>=0.18.0
python-frontmatter>=1.0.0
pyyaml>=6.0
